[Rename] Finalizing
**Major**:
- `stream_data` to `event_data`
- some `stream_publisher` to `event_publisher`

All tests are passing. Direct usage of hatchling including MCP servers also worked.

**Minor**:
- Missed renaming some import paths `*core.chat*` to `*ui*` for the CLI commands
- Fixed a wrong renaming that produced the import paths `*Hatchling.hatchling.*`
