+++ hatchling/core/llm/providers/base.py
@@ -9,7 +9,7 @@ from abc import ABC, abstractmethod
 from typing import Dict, List, Any, Optional
 
 from hatchling.core.llm.event_system import EventPublisher
-from Hatchling.hatchling.mcp_utils.mcp_tool_lifecycle_subscriber import ToolLifecycleSubscriber
+from hatchling.mcp_utils.mcp_tool_lifecycle_subscriber import ToolLifecycleSubscriber
 from hatchling.core.llm.event_system.event_data import Event
 from hatchling.core.llm.data_structures import ToolCallParsedResult
 from hatchling.config.settings import AppSettings
