+++ tests/integration_test_openai.py
@@ -35,7 +35,7 @@ from tests.test_decorators import slow_test, requires_api_key, integration_test,
 from hatchling.config.openai_settings import OpenAISettings
 from hatchling.config.settings import AppSettings
 from hatchling.config.llm_settings import ELLMProvider
-from Hatchling.hatchling.mcp_utils.mcp_tool_lifecycle_subscriber import ToolLifecycleSubscriber
+from hatchling.mcp_utils.mcp_tool_lifecycle_subscriber import ToolLifecycleSubscriber
 from hatchling.core.llm.providers.registry import ProviderRegistry
 from hatchling.mcp_utils.mcp_tool_data import MCPToolInfo, MCPToolStatus, MCPToolStatusReason
 from hatchling.core.llm.event_system import (
@@ -47,7 +47,7 @@ from hatchling.core.llm.event_system import (
     EventType,
     Event
 )
-from Hatchling.hatchling.mcp_utils.mcp_tool_lifecycle_subscriber import ToolLifecycleSubscriber
+from hatchling.mcp_utils.mcp_tool_lifecycle_subscriber import ToolLifecycleSubscriber
 from hatchling.core.llm.providers.openai_provider import OpenAIProvider
 
 logger = logging.getLogger("integration_test_openai")
