+++ tests/regression_test_enhanced_tool_execution.py
@@ -52,7 +52,7 @@ class TestToolExecutionRegression(unittest.TestCase):
         self.assertTrue(hasattr(self.tool_execution, 'current_tool_call_iteration'))
         self.assertTrue(hasattr(self.tool_execution, 'tool_call_start_time'))
         self.assertTrue(hasattr(self.tool_execution, 'root_tool_query'))
-        self.assertTrue(hasattr(self.tool_execution, 'stream_publisher'))
+        self.assertTrue(hasattr(self.tool_execution, 'event_publisher'))
     
     @regression_test
     def test_mcp_manager_integration_still_works(self):
@@ -162,11 +162,11 @@ class TestToolExecutionRegression(unittest.TestCase):
     @regression_test
     def test_event_publisher_integration_still_works(self):
         """Test that stream publisher integration still works."""
-        # The new interface should have stream_publisher property
-        self.assertTrue(hasattr(self.tool_execution, 'stream_publisher'))
+        # The new interface should have event_publisher property
+        self.assertTrue(hasattr(self.tool_execution, 'event_publisher'))
         
         # Should be able to access the publisher
-        publisher = self.tool_execution.stream_publisher
+        publisher = self.tool_execution.event_publisher
         self.assertIsNotNone(publisher)
         
         # Should have the necessary methods for event publishing
@@ -192,7 +192,7 @@ class TestToolExecutionRegression(unittest.TestCase):
         self.assertTrue(hasattr(self.tool_execution, '_event_publisher'))
         
         # Test that events can be published (basic functionality test)
-        publisher = self.tool_execution.stream_publisher
+        publisher = self.tool_execution.event_publisher
         
         # Should not raise an exception when publishing events
         try:
@@ -219,7 +219,7 @@ class TestToolExecutionRegression(unittest.TestCase):
     def test_new_event_publisher_doesnt_break_existing_functionality(self):
         """Test that the new EventPublisher doesn't interfere with existing functionality."""
         # Verify the stream publisher exists (new functionality)
-        self.assertTrue(hasattr(self.tool_execution, 'stream_publisher'))
+        self.assertTrue(hasattr(self.tool_execution, 'event_publisher'))
         self.assertTrue(hasattr(self.tool_execution, '_event_publisher'))
         
         # Verify that having a stream publisher doesn't break basic operations
